add rarisdk as git submodule and import addresses from there
