--- scripts/sdk/Fuse/contracts/addresses/arbitrumRinkeby.ts
@@ -1,112 +1,0 @@
-export const arbitrumRinkeby = {
-  // Fundamentals
-  FUSE_POOL_DIRECTORY_CONTRACT_ADDRESS:
-    "0x440253a301b4Fc749a24E653fDb11B742e621892",
-  FUSE_SAFE_LIQUIDATOR_CONTRACT_ADDRESS:
-    "0x53cC546053a09431c9d1A2b288471680d15000a4",
-  FUSE_FEE_DISTRIBUTOR_CONTRACT_ADDRESS:
-    "0xEf562dd508f3dBBa4B4ad62a869627D379692557",
-  FUSE_POOL_LENS_CONTRACT_ADDRESS: "0xBF29f83D9a0EC1d9eC7281381E04750b817feA4E",
-  FUSE_POOL_LENS_SECONDARY_CONTRACT_ADDRESS:
-    "0xb2a4B68f5126F1624b900f5d637Bd15968d7a119",
-
-  // CEther and CERC20
-  COMPTROLLER_IMPLEMENTATION_CONTRACT_ADDRESS:
-    "0xCE8322B3C5363c0eb98A490bF42d4e5F4670F354",
-  CETHER_DELEGATE_CONTRACT_ADDRESS:
-    "0xB50eF61B568c10894E1B3DE308a31B8E75e9B4ff",
-  CERC20_DELEGATE_CONTRACT_ADDRESS:
-    "0xaa029B1A8050979934093cf76eC9E8c13fE16CBf",
-
-  // Oracles
-  MASTER_PRICE_ORACLE_IMPLEMENTATION_CONTRACT_ADDRESS:
-    "0x6AB2497DB9F36ebD69d2F9F13329A27Fe070af09",
-  INITIALIZABLE_CLONES_CONTRACT_ADDRESS:
-    "0x374D551f0584107d6B4D2B061E09436c0FA0C098",
-
-  PUBLIC_PRICE_ORACLE_CONTRACT_ADDRESSES: {
-    MasterPriceOracle: "0xc5a8c01509c1972895dfb23f6bdf304365cbcce0", // Rari DAO MasterPriceOracle actual
-    ChainlinkPriceOracleV2: "0xCc5F01efd5647C77861Ce15C923731f9e27E3132",
-    FixedEurPriceOracle: "0x05FBd6E256d8f481F4fc3Ffb113e076d4919153f",
-    FixedETHPriceOracle: "0xD7d2d97C61AfB3045039584Dc654B926694083C5",
-  },
-
-  PRICE_ORACLE_RUNTIME_BYTECODE_HASHES: {
-    MasterPriceOracle:
-      "0x9715c015e495c902550ebbb41e79ac57130863255b4715a7d73fc38c0d578acf",
-    ChainlinkPriceOracleV2:
-      "0xc43e378a0381e1cbf6b3d5f2516e2dbc8265f973c107c62253f08d5b0d53bfda",
-    FixedEurPriceOracle:
-      "0x678dbe9f2399a44e89edc934dc17f6d4ee7004d9cbcee83c0fa0ef43de924b84",
-    FixedETHPriceOracle:
-      "0xcb669c93632a1c991adced5f4d97202aa219fab3d5d86ebd28f4f62ad7aa6cb3",
-  },
-
-  ORACLES: ["MasterPriceOracle", "ChainlinkPriceOracleV2"],
-
-  oracles: {
-    MasterPriceOracle: {
-      address: "0xc5a8c01509c1972895dfb23f6bdf304365cbcce0",
-      bytecodeHash:
-        "0x9715c015e495c902550ebbb41e79ac57130863255b4715a7d73fc38c0d578acf",
-      deployable: false,
-      oldVersions: {},
-    },
-    ChainlinkPriceOracleV2: {
-      address: "0xCc5F01efd5647C77861Ce15C923731f9e27E3132",
-      bytecodeHash:
-        "0xc43e378a0381e1cbf6b3d5f2516e2dbc8265f973c107c62253f08d5b0d53bfda",
-      deployable: false,
-      oldVersions: {},
-    },
-    FixedEurPriceOracle: {
-      address: "0x817158553F4391B0d53d242fC332f2eF82463e2a",
-      bytecodeHash:
-        "0x678dbe9f2399a44e89edc934dc17f6d4ee7004d9cbcee83c0fa0ef43de924b84",
-      deployable: false,
-      oldVersions: {},
-    },
-    FixedETHPriceOracle: {
-      address: "0xD7d2d97C61AfB3045039584Dc654B926694083C5",
-      bytecodeHash:
-        "0xcb669c93632a1c991adced5f4d97202aa219fab3d5d86ebd28f4f62ad7aa6cb3",
-      deployable: false,
-      oldVersions: {},
-    },
-  },
-
-  DEPLOYABLE_ORACLES: [],
-
-  // // UNI-V2 Oracles (Sushiswap for arbitrum)
-  UNISWAP_V2_FACTORY_ADDRESS: "0xc35dadb65012ec5796536bd9864ed8773abc74c4",
-  UNISWAP_V2_PAIR_INIT_CODE_HASH: "", // Todo - remove bc only used in checkForCErc20PriceFeed which is now deleted
-
-  UNISWAP_TWAP_PRICE_ORACLE_ROOT_CONTRACT_ADDRESS: "", // Unused - only on ETH
-  UNISWAP_TWAP_PRICE_ORACLE_V2_ROOT_CONTRACT_ADDRESS:
-    "0x1AC1EC4030D3EBaeae7BE8d5786a8a7Bca071A06",
-  UNISWAP_TWAP_PRICE_ORACLE_V2_FACTORY_CONTRACT_ADDRESS:
-    "0xa18C88f04E46Caa1541c90E8eB1a25F9F45B8C8a",
-
-  // // UNI-V3 Oracles
-  UNISWAP_V3_FACTORY_ADDRESS: "",
-  UNISWAP_V3_TWAP_PRICE_ORACLE_V2_FACTORY_CONTRACT_ADDRESS:
-    "0x2B88F7F53649EfAE653e53A204B4618EdE06Dd29",
-
-  // IRM
-  PUBLIC_INTEREST_RATE_MODEL_CONTRACT_ADDRESSES: {
-    JumpRateModel_Cream_Stables_Majors:
-      "0x1cE7281CE601d06cf5f47A3023b80ebd73E3B35D",
-    JumpRateModel_Cream_Major: "0x8dbf1250c805fc2ed29fc0d3aed31ec69a928ffe",
-    JumpRateModel_Cream_Gov: "0x46c54c7D214117c79f2f6F368549776F00c0a6c4",
-  },
-
-  // Tokens / ETC
-  WETH_ADDRESS: "0x7ceB23fD6bC0adD59E62ac25578270cFf1b9f619",
-  REWARDS_DISTRIBUTOR_DELEGATE_CONTRACT_ADDRESS: "",
-
-  // Legacy
-  COINBASE_PRO_REPORTER_ADDRESS: "0xfCEAdAFab14d46e20144F48824d0C09B1a03F2BC",
-  DAI_POT: "",
-  DAI_JUG: "",
-  OPEN_ORACLE_PRICE_DATA_CONTRACT_ADDRESS: "",
-};
